chore: extract internal gh package
