+++ internal/gh/doc.go
@@ -1,0 +1,2 @@
+// Package gh provides primitives for invoking the GitHub CLI.
+package gh
